# openapi/hub.openapi.json
{
  "components": {
    "schemas": {
...
      "FleetSpendView": {
        "description": "The fleet's usage/cost total since ``since``. ``cost_partial`` carries the\nlower-bound + PARTIAL contract on ``cost_usd`` \u2014 see\n:class:`~blizzard.hub.domain.work.UsageTotal` for the one canonical statement of\nit, which this view's fields mirror verbatim.",
        "properties": {
          "cache_create_tokens": {
            "title": "Cache Create Tokens",
            "type": "integer"
          },
          "cache_read_tokens": {
            "title": "Cache Read Tokens",
            "type": "integer"
          },
          "cost_partial": {
            "title": "Cost Partial",
            "type": "boolean"
          },
          "cost_usd": {
            "title": "Cost Usd",
            "type": "number"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "since": {
            "title": "Since",
            "type": "string"
          }
        },
        "required": [
          "since",
          "input_tokens",
          "output_tokens",
          "cache_read_tokens",
          "cache_create_tokens",
          "cost_usd",
          "cost_partial"
        ],
        "title": "FleetSpendView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunnerListResponse": {
        "description": "The fleet registry \u2014 every registered runner with its liveness.",
        "properties": {
          "runners": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RunnerView"
            },
            "title": "Runners",
            "type": "array"
          }
        },
        "title": "RunnerListResponse",
        "type": "object"
      },
      "RunnerPauseRequest": {
        "description": "Set a runner's pause brake \u2014 records who flipped it.",
        "properties": {
          "by": {
            "default": "operator",
            "title": "By",
            "type": "string"
          }
        },
        "title": "RunnerPauseRequest",
        "type": "object"
      },
      "RunnerRegistrationRequest": {
        "description": "Register a runner into the fleet \u2014 runner id + workspace binding.",
        "properties": {
          "runner_id": {
            "title": "Runner Id",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "runner_id",
          "workspace_id"
        ],
        "title": "RunnerRegistrationRequest",
        "type": "object"
      },
      "RunnerRegistrationResponse": {
        "description": "The registered runner's id, and whether this call first created its row.",
        "properties": {
          "first_registration": {
            "title": "First Registration",
            "type": "boolean"
          },
          "runner_id": {
            "title": "Runner Id",
            "type": "string"
          }
        },
        "required": [
          "runner_id",
          "first_registration"
        ],
        "title": "RunnerRegistrationResponse",
        "type": "object"
      },
      "RunnerView": {
        "description": "One fleet-registry row \u2014 derived liveness and both brakes.\n\nA runner can be paused by two different parties for two different reasons, so the two\nare reported separately rather than collapsed into one ``paused`` (issue #43): the\nboard shows *which*. A reader that wants \"is it claiming?\" ORs them; since issue #45\nthe two diverge past claiming \u2014 ``hub_paused`` keeps its claims-only meaning, while\n``locally_paused`` alone answers \"is it spawning anything at all?\".",
        "properties": {
          "hub_paused": {
            "title": "Hub Paused",
            "type": "boolean"
          },
          "last_seen_at": {
            "title": "Last Seen At",
            "type": "string"
          },
          "locally_paused": {
            "default": false,
            "title": "Locally Paused",
            "type": "boolean"
          },
          "locally_paused_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locally Paused By"
          },
          "locally_paused_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locally Paused Reason"
          },
          "online": {
            "title": "Online",
            "type": "boolean"
          },
          "registered_at": {
            "title": "Registered At",
            "type": "string"
          },
          "runner_id": {
            "title": "Runner Id",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "runner_id",
          "workspace_id",
          "registered_at",
          "last_seen_at",
          "online",
          "hub_paused"
        ],
        "title": "RunnerView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "blizzard-hub",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/fleet/spend": {
      "get": {
        "description": "The fleet's total usage/cost since ``since`` (an ISO-8601 instant) \u2014 summed\nover every usage fact recorded at or after it, across every chunk.",
        "operationId": "fleet_spend_api_fleet_spend_get",
        "parameters": [
          {
            "in": "query",
            "name": "since",
            "required": true,
            "schema": {
              "title": "Since",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetSpendView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fleet Spend",
        "tags": [
          "fleet"
        ]
      }
    },
...
    "/api/runners": {
      "get": {
        "description": "The fleet registry \u2014 every runner with derived liveness + paused state.",
        "operationId": "list_runners_api_runners_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunnerListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Runners",
        "tags": [
          "runners"
        ]
      },
      "post": {
        "description": "Register a runner \u2014 runner id + workspace binding; idempotent upsert.",
        "operationId": "register_runner_api_runners_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunnerRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunnerRegistrationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Runner",
        "tags": [
          "runners"
        ]
      }
    },
    "/api/runners/{runner_id}": {
      "get": {
        "description": "One runner's declarative state \u2014 the runner's own pull read.",
        "operationId": "get_runner_api_runners__runner_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "runner_id",
            "required": true,
            "schema": {
              "title": "Runner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunnerView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Runner",
        "tags": [
          "runners"
        ]
      }
    },
    "/api/runners/{runner_id}/heartbeats": {
      "post": {
        "description": "Refresh a runner's liveness \u2014 the slow runner-level heartbeat. Returns 204.",
        "operationId": "heartbeat_runner_api_runners__runner_id__heartbeats_post",
        "parameters": [
          {
            "in": "path",
            "name": "runner_id",
            "required": true,
            "schema": {
              "title": "Runner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Heartbeat Runner",
        "tags": [
          "runners"
        ]
      }
    },
    "/api/runners/{runner_id}/pause": {
      "post": {
        "description": "Set a runner's pause brake \u2014 no new claims; in-flight chunks run on.",
        "operationId": "pause_runner_api_runners__runner_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "runner_id",
            "required": true,
            "schema": {
              "title": "Runner Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunnerPauseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunnerView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pause Runner",
        "tags": [
          "runners"
        ]
      }
    },
    "/api/runners/{runner_id}/resume": {
      "post": {
        "description": "Clear a runner's pause brake \u2014 it resumes claiming on its next pull.",
        "operationId": "resume_runner_api_runners__runner_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "runner_id",
            "required": true,
            "schema": {
              "title": "Runner Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunnerPauseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunnerView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Runner",
        "tags": [
          "runners"
        ]
      }
    }
  }
}